fixed python and C++ code for new framework
